% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CS3_main_functions.R
\name{combined_pessc_scores}
\alias{combined_pessc_scores}
\title{Generate average component-sample similarity score.}
\usage{
combined_pessc_scores(per_eem_ssc_res, LoQ_threshold = TRUE)
}
\arguments{
\item{per_eem_ssc_res}{an output from per_eem_ssc}

\item{LoQ_threshold}{logical TRUE/FALSE. If TRUE, the function will only use scores for samples that exceed the LoQ. This will have been calculated previously if an MQL EEM object was supplied to the per_eem_ssc function.}
}
\description{
Produce a simple average of PARAFAC component-sample similarity scores, with
option for limit of quantification (LoQ) threshold.
}
