% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{create_MDL_eem}
\alias{create_MDL_eem}
\title{Create an EEM object representing the Long-Term Method Detection Limit LT-MDL}
\usage{
create_MDL_eem(
  blank_eemlist,
  remove_gamma_spikes = TRUE,
  excise_scatter = c(TRUE, TRUE, TRUE, TRUE),
  scatter_widths = c(12, 12, 10, 10),
  method = "Hansen"
)
}
\arguments{
\item{blank_eemlist}{an eemlist compliant with the eem/eemR/staRdom framework, containing a set of ultrapure water blank EEMs created using identical measurement parameters.}

\item{remove_gamma_spikes}{uses forecast::tsclean to treat each emission scan in the LT-MDL as a time series and remove outliers. This gets rid of single-point scatter artefacts caused by gamma spikes.}

\item{excise_scatter}{four logical inputs, e.g. c(TRUE,TRUE,TRUE,TRUE). Passed to staRdom::eem_rem_scat. See ?staRdom::eem_rem_scat for more information.}

\item{scatter_widths}{four numeric inputs, specifying the width, in nm, of scatter excissions. See ?staRdom::eem_rem_scat.}

\item{method}{character vector, one of either 'Hansen' or 'Thomsen' corresponding to the two supproted methods for LT-MDL calculation.}
}
\description{
The LT-MDL quantifies a given fluorometer's ability to detect signals
at specific wavelengths by combining the response of a set of ultrapure water blanks
run using the same parameters as the method of interest. Two options are available for
calculating the LT-MDL: the average of a set of ultrapure water blanks plus 3* the standard
deviation of the blanks (following Hansen et al., 2018), or simply 3* the standard deviation of
the blanks (following Thomsen et al., 2003).
}
