% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{eemlist_average_int}
\alias{eemlist_average_int}
\title{Average a set of EEMs.}
\usage{
eemlist_average_int(eemlist)
}
\arguments{
\item{eemlist}{A list of EEMs, compliant with the eemR/staRdom framework.}
}
\description{
Average the EEMs within an eemlist. Borrowed from eemUtils.
}
