% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sg_terms_deftab}
\alias{sg_terms_deftab}
\title{Default terms for Savitzky-Golay smoothing}
\format{
A data frame with 2 rows and 4 variables:
\describe{
\item{p}{polynomial order}
\item{n}{width of filter window, in nm}
...
}
}
\usage{
sg_terms_deftab
}
\description{
A table containing the default values for the signal::sgolayfilt() function.
}
\keyword{datasets}
