% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper_functions.R
\name{ssc_more_int}
\alias{ssc_more_int}
\title{Get SSC along with alpha and beta penalty terms}
\usage{
ssc_more_int(mat1, mat2, tcc = FALSE)
}
\arguments{
\item{mat1}{a matrix}

\item{mat2}{a matrix}

\item{tcc}{TRUE/FALSE to return only TCC value instead of SSC, alpha and beta.}
}
\description{
The shift- and shape sensitive congruence (SSC) was developed by Wunsch et al., 2019 as an improvement
upon the TCC metric. It incorporates two penalty terms, alpha and beta, to account for differences in the wavelength
peak position and area. This function adds these terms to the data frame returned by staRdom::ssc(). Code from https://github.com/MRPHarris/eemUtils.
}
