% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CS3_main_function.R
\name{per_eem_ssc}
\alias{per_eem_ssc}
\title{Compare PARAFAC and EEM spectra with spectral similarity metrics}
\usage{
per_eem_ssc(
  pfmodel,
  eemlist,
  comp,
  tcc = FALSE,
  modified_metrics = TRUE,
  spectral_correct = "all",
  interp_1nm = "all",
  smooth_sg = "all",
  complete_peak = "ex",
  SSC_trim_method = "mod_alpha",
  verbose = FALSE,
  denormalise_residuals = FALSE,
  exclude_negative_residuals = FALSE
)
}
\arguments{
\item{pfmodel}{a PARAFAC model object. An output from staRdom::eem_parafac()}

\item{eemlist}{a group of EEMs compliant with the staRdom/EEM/eemR package framework}

\item{comp}{numeric, singular - the target component for comparisons}

\item{tcc}{TRUE/FALSE to extract only TCC rather than SSC}

\item{modified_metrics}{TRUE/FALSE to add the modified (ex + em combined) TCC and SSC metrics after Parr et al., 2014.}

\item{spectral_correct}{either NULL, 'all', 'ex', or 'em'. Subtract the loadings of other components from the raw sample spectra of the type specified}

\item{interp_1nm}{either NULL, 'all', 'ex', or 'em'.  Interpolate spectra to 1nm bandwidths. Recommended. Applied after spectral correction}

\item{smooth_sg}{either NULL, 'all', 'ex', or 'em'. Applies a Savitzky-Golay filter to the data prior to metric calculation. Only recommended if interpolation is performed. Uses signal::sgolay(). Default values are a 2nd order polynomial, n = 21 for emission spectra and n = 11 for excitation spectra}

\item{complete_peak}{either NULL, 'all', 'ex', or 'em'. Use a gradient detection method to identify incomplete peaks in spectra, and prevent them from interfering with the SSC peak position calculation. Set options with the parameter 'SSC_trim_method'.}

\item{SSC_trim_method}{if !is.null(complete_peak), choose one of either 'mod_alpha' to alter the peak position penalty term, or 'trim_spectra' to remove the incomplete peak entirely}

\item{verbose}{TRUE/FALSE to return various messages during the function's opperation. Useful for error checking or to keep track of how things are proceeding. Only used for spectral correction.}

\item{denormalise_residuals}{TRUE/FALSE to denormalise residuals using the max fluorescence value of the supplied eemlist. Default to FALSE.}

\item{exclude_negative_residuals}{TRUE/FALSE to set negative values to 0 during component overlap correction. Setting this to TRUE may inflate similarity scores in component:sample comparisons where the model has over-fitted the component.}

\item{terms}{TRUE/FALSE to extract the alpha and beta penalty term values alongside SSC}
}
\description{
Extract the spectra from a PARAFAC model (B and C mode loadings) and compare them with underlying
sample data. Various options exist for corrections, including adjusting excitation spectra so that SSC might
be applied rather than just TCC.
}
